// src/models/Api.ts
export interface SwapiPerson {
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
}

export interface Cordenadas {
  ip: string;
  city: string;
  region: string;
  country: string;
  latitude: string;
  longitude: string;
}